+++ src/store/store.ts
@@ -1,10 +1,29 @@
-import { createStore, applyMiddleware, Store } from "redux"
+import { createStore, applyMiddleware, Store, Action } from "redux"
 import thunk from "redux-thunk"
-import { reducer } from "src/context/reducer";
+import { Config, createStateSyncMiddleware } from "redux-state-sync";
+import { persistStore } from 'redux-persist'
+import { PersistPartial } from 'redux-persist/lib/persistReducer';
+import persistingReducer from "./persistingReducer";
 
-// TODO: reimplement this class properly
-export const store: Store<any, any> & {
-    dispatch: DispatchType
-} = createStore(reducer, applyMiddleware(thunk));
+// The redux state sync configuration
+const STATE_SYNC_CONFIG: Config = {
+    channel: 'react_todo_list_channel',
+};
 
-export default store;
+// All the redux middlewares
+const storeMiddlewares = [
+    createStateSyncMiddleware(STATE_SYNC_CONFIG),
+    thunk,
+];
+
+export default () => {
+    // Create the redux store
+    const store: Store<PersistPartial, Action<any>> & {
+        dispatch: DispatchType
+    } = createStore(persistingReducer, applyMiddleware(...storeMiddlewares));
+
+    // And it's corresponding persisto
+    const persistor = persistStore(store);
+
+    return { store, persistor }
+};